co/backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
python-dotenv==1.1.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-generativeai==0.8.5
requests==2.32.3
PyJWT==2.10.1
